Keep image zone rules cache separate #2637
* Before this patch,
the image would contain zone rules cache from the JVM computing the image.
* Side effects here,
such as class or source cache lookups,
were affecting its contents and leading to error when writing image.
* With this change,
we guarantee that the zone rules cache in the image is a brand new instance,
separate from the one in the JVM computing the image.