Fixes inlined code's cache paths for DWARF generation
Since inlined methods may be defined in a different file than the one
defining a compiled method, the cache paths of all the source files
contributing code to the compiled method need to be set as the
compilation directory to make gdb work.

E.g. If:
1. `foo` invokes `bar` and `baz`
2. `bar` and `baz` get inlined in `foo`
3. `foo`'s source file is cached in `sources/src`
4. `bar`'s source file is cached in `sources/jdk`
5. `baz`'s source file is cached in `sources/graal`

then the compilation directory in `foo`'s compilation unit needs to be
set to `sources/src:sources/jdk:sources/graal`